requirements.txt
aiogram>=3.5.0
aiohttp>=3.9.5
APScheduler>=3.10.4
pytz>=2024.1
aiosqlite>=0.20.0
python-dotenv>=1.0.1